third_party/WebKit/LayoutTests/fast/dom/NodeList/nodelist-item-with-index-expected.txt

This tests that items in a NodeList can be retrieved by index.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

text1
text2


PASS div0 == div0s is true
PASS div0 == div0s_ is false
PASS div0 == div1 is false
PASS div0s == div0s_ is false
PASS div0s == div1 is false
PASS div0s_ == div1 is false
PASS successfullyParsed is true

TEST COMPLETE
